add width & height to bimdata-dropdown-list (#119)
current state : the minimum width and height were assigned to the first child of bimdata-dropdown-list. When using the component, the parent had to be assigned a width so that the dropdown list was not wider than its parent.

new state : the same height and width assigned to the first child, are also assigned to the parent.